refactor(cat): split context-string-builder into section renderers (605→95)
context-string-builder.ts was 605 LOC — the builder + all 17 inline context
sections in one function. Extracted each section into a pure render*() function
in context-sections.ts; buildFullContextString is now a 95-LOC composition
(render in priority order → budget → wrap).

Output is BYTE-IDENTICAL (it feeds the Cat prompt every turn): added
context-string-builder.snapshot.test.ts with an all-sections fixture + fixed
system time, captured the baseline from the pre-refactor code, and the refactor
matches it exactly. tsc 0; all existing runtime-context tests still pass.

context-sections.ts is 572 LOC but it's a flat list of independent, pure
renderers (high cohesion, no coupling) — not a tangled god-file.
